+++ cinelerra/edl.h
@@ -205,7 +205,7 @@ public:
 	void calculate_conformed_dimensions(int single_channel, float &w, float &h);
 // Get the total output size scaled to aspect ratio
 	void output_dimensions_scaled(int &w, int &h);
-	float get_aspect_ratio();
+	double get_aspect_ratio();
 
 // Clips
 	ArrayList<EDL*> clips;
